+++ libcontainerd/supervisor/remote_daemon_linux.go
@@ -2,33 +2,11 @@ package supervisor // import "github.com/docker/docker/libcontainerd/supervisor"
 
 import (
 	"os"
-	"path/filepath"
 	"syscall"
 	"time"
 
-	"github.com/containerd/containerd/defaults"
 	"github.com/docker/docker/pkg/process"
 )
-
-const (
-	sockFile      = "containerd.sock"
-	debugSockFile = "containerd-debug.sock"
-)
-
-func (r *remote) setDefaults() {
-	if r.GRPC.Address == "" {
-		r.GRPC.Address = filepath.Join(r.stateDir, sockFile)
-	}
-	if r.GRPC.MaxRecvMsgSize == 0 {
-		r.GRPC.MaxRecvMsgSize = defaults.DefaultMaxRecvMsgSize
-	}
-	if r.GRPC.MaxSendMsgSize == 0 {
-		r.GRPC.MaxSendMsgSize = defaults.DefaultMaxSendMsgSize
-	}
-	if r.Debug.Address == "" {
-		r.Debug.Address = filepath.Join(r.stateDir, debugSockFile)
-	}
-}
 
 func (r *remote) stopDaemon() {
 	// Ask the daemon to quit
