+++ libcontainerd/supervisor/remote_daemon_windows.go
@@ -6,20 +6,6 @@ import (
 	"github.com/docker/docker/pkg/process"
 )
 
-const (
-	grpcPipeName  = `\\.\pipe\containerd-containerd`
-	debugPipeName = `\\.\pipe\containerd-debug`
-)
-
-func (r *remote) setDefaults() {
-	if r.GRPC.Address == "" {
-		r.GRPC.Address = grpcPipeName
-	}
-	if r.Debug.Address == "" {
-		r.Debug.Address = debugPipeName
-	}
-}
-
 func (r *remote) stopDaemon() {
 	p, err := os.FindProcess(r.daemonPid)
 	if err != nil {
